Import count reducer
